# optimize slow-query list performance